Initial commit with project boilerplate.
This commit contains the initial boilerplate for
an Elixir API client. Hopefully the pattern used
will be fine for all of the bandwidth resources.
Higher level abstractions like actually encoding
and decoding of bandwidth models will be done later.

For now this project is only concerned with providing
an abstraction for all REST resources available
through the API.