Cleaned up resource methods and response type.
To be idiomatic about error handling the return
type of any request functions in the client module
are in tuple form. Additionally, the success type
of {:ok, result} is {status_code, data, headers}.

Since the top level API has changed the version has
been bumped to 1.0.0.

Also introduced are the beginnings of a testing
strategy for the project. The Client is where a
bulk of testing will be done. The tests in this
commit could probably be made simpler.

A few extra resources were added in this change.

Future testing should be very easy since all resources
follow the same basic pattern.